moved constants to a common location and setting scheduling on main app
