+++ src/main/java/com/robertabela/rss/Constants.java
