+++ src/main/java/com/robertabela/rss/RssFeedApplication.java
@@ -2,8 +2,10 @@ package com.robertabela.rss;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
+@EnableScheduling
 public class RssFeedApplication {
 
 	public static void main(String[] args) {
